+++ packages/enhanced/src/lib/container/ModuleFederationPlugin.ts
@@ -11,6 +11,7 @@ import {
   composeKeyWithSeparator,
   type moduleFederationPlugin,
 } from '@module-federation/sdk';
+import { PrefetchPlugin } from '@module-federation/data-prefetch/cli';
 import { normalizeWebpackPath } from '@module-federation/sdk/normalize-webpack-path';
 import type { Compiler, WebpackPluginInstance } from 'webpack';
 import SharePlugin from '../sharing/SharePlugin';
@@ -73,6 +74,7 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
     if (options.dts !== false) {
       new DtsPlugin(options).apply(compiler);
     }
+    new PrefetchPlugin(options).apply(compiler);
 
     new FederationRuntimePlugin(options).apply(compiler);