+++ packages/runtime/src/core.ts
@@ -1,5 +1,8 @@
-import type { CreateScriptHookReturn } from '@module-federation/sdk';
 import { isBrowserEnv } from '@module-federation/sdk';
+import type {
+  CreateScriptHookReturn,
+  ModuleInfo,
+} from '@module-federation/sdk';
 import {
   Options,
   PreloadRemoteArgs,
@@ -67,6 +70,8 @@ export class FederationHost {
       remoteInfo: RemoteInfo;
       remoteEntryExports: RemoteEntryExports;
       origin: FederationHost;
+      id: string;
+      remoteSnapshot?: ModuleInfo;
     }>('initContainer'),
   });
   version: string = __VERSION__;