+++ apps/router-demo/router-host-v5-2200/CHANGELOG.md
@@ -1,5 +1,12 @@
 # host-v5
 
+## 1.0.24
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.4
+- @module-federation/bridge-react@0.6.4
+
 ## 1.0.23
 
 ### Patch Changes